Move pagestream api to libs/pageserver_api (#2698)
